fix: must provide outdir param to snowdrop generator
Fixes #50. outdir is needed because it is used by the generator to infer the
component name (which is wrong). We meed to wait until Validate to be able to
finish the generator URL because Name is not set in Complete (it's set in the
generic CreateOptions.Complete after the child version of Complete is called).