feat: initial implementation of halkyon descriptor support
When hal runs to attempt to create an entity, it will gather all locally known
entities of that type by parsing any existing halkyon descriptors. hal will look
for descriptors as follows: halkyon.(yml|yaml) files in the current directory,
dekorate-generated descriptor for the current directory and look for the same
files in each of the current directory's child directories, not looking further
than one level deep. When descriptors are detected, the user is asked whether they
want to operate on the detected entities.